+++ src/test/unit/mocks/mocked-requests/other-proceedings-mock.ts
@@ -1,4 +1,4 @@
-import { mockRequest } from '../../../../test/unit/utils/mockRequest';
+import { mockRequest } from '../../utils/mockRequest';
 
 export const otherProceedingsMockData = mockRequest({
   query: {
