+++ src/test/unit/mocks/mocked-requests/respondent-details-mock.ts
@@ -1,4 +1,4 @@
-import { mockRequest } from '../../../../test/unit/utils/mockRequest';
+import { mockRequest } from '../../utils/mockRequest';
 
 export const respondentMockData = mockRequest({
   params: {
